decouple sharding config from sharding package root
