+++ usecases/scaler/scaler.go
@@ -21,6 +21,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/weaviate/weaviate/entities/backup"
 	"github.com/weaviate/weaviate/usecases/sharding"
+	"github.com/weaviate/weaviate/usecases/sharding/config"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -97,7 +98,7 @@ func (s *Scaler) SetSchemaManager(sm SchemaManager) {
 // make sure to broadcast that state to all nodes as part of the "update"
 // transaction.
 func (s *Scaler) Scale(ctx context.Context, className string,
-	updated sharding.Config, prevReplFactor, newReplFactor int64,
+	updated config.Config, prevReplFactor, newReplFactor int64,
 ) (*sharding.State, error) {
 	// First identify what the sharding state was before this change. This is
 	// mainly to be able to compare the diff later, so we know where we need to
@@ -123,7 +124,7 @@ func (s *Scaler) Scale(ctx context.Context, className string,
 // * It pushes locally existing shards to new replicas
 // * It delegates replication of remote shards to owner nodes
 func (s *Scaler) scaleOut(ctx context.Context, className string, ssBefore *sharding.State,
-	updated sharding.Config, replFactor int64,
+	updated config.Config, replFactor int64,
 ) (*sharding.State, error) {
 	// Create a deep copy of the old sharding state, so we can start building the
 	// updated state. Because this is a deep copy we don't risk leaking our
@@ -214,7 +215,7 @@ func (s *Scaler) LocalScaleOut(ctx context.Context,
 }
 
 func (s *Scaler) scaleIn(ctx context.Context, className string,
-	updated sharding.Config,
+	updated config.Config,
 ) (*sharding.State, error) {
 	return nil, errors.Errorf("scaling in not supported yet")
 }
