+++ usecases/sharding/config/config_test.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package sharding
+package config
 
 import (
 	"encoding/json"
