+++ usecases/sharding/config/config_update.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package sharding
+package config
 
 import (
 	"fmt"
